Introduce PjRtLayout and change existing APIs to use it.
This introduces a new class, PjRtLayout, to replace direct uses of
xla::Layout in the PJRT layer and above. The goal is limit the API
surface area outside of XLA -- xla::Layout contains many fields that
are used inside the compiler but aren't meaningful higher in the
stack.

This change also modifies PjRtExecutable::GetParameterLayouts() and
PjRtExecutable::GetOutputLayouts() to return PjRtLayouts instead of
xla::Layouts.

This change also introduces a PjRtLayout subclass, PjRtXlaLayout,
which is backed by an xla::Layout. This is used by default
implementations. This is still a bit sketchy because the underlying
xla::Layout may contain information that will be ignored by higher
layers of the stack, but at least the public interface makes that more
clear.
PiperOrigin-RevId: 588247567