[XLA] Fix incorrect results of raising large complex numbers to complex powers.
I replaced a naive algorithm that computes `sqrt(a*a+b*b)` with the existing algorithm for complex `Abs`. The naive computation overflowed as soon as `a*a` or `b*b` overflowed.

I did not tweak the existing computation for `Abs` though we could increase the precision if we implement:

CARLOS F. BORGES - AN IMPROVED ALGORITHM FOR HYPOT(A,B): https://arxiv.org/pdf/1904.09481.pdf

Fixes openxla/xla#7440

Fixes google/jax#18749

PiperOrigin-RevId: 592851796